split up mpath subcommands into separate files
